Fix save behavior
In the previous commit message, I described some unintended behavior
that occurs when a user saves a PDF post. When a user attempted to
update the PDF post but *did not* provide a PDF file, any previously
uploaded file would be disassociated from the post.

This commit fixes that issue. Now, if a file is not supplied, no action
is taken. Additionally, upload attempts that contain incorrect/
unsupported filetypes will prompt an error message.

This functionality was not introduced via AJAX/JavaScript (as hinted in
the previous commit message). Instead, all validation and persistence
is handled on server-side.

Lastly, this commit cleans up some of the rough edges that were present
in the prior commit. This comit provides a better object-oriented
design than the previous iteration, and removes many unused files.