Include a customized template for search results
Previously, I saved a copy of one of the theme's files in a `backups`
directory within the plugin to safeguard against overwrites that could
affect the way that PDFs look in the search results.

A better strategy is to include a default template that can be pasted
into the theme when necessary.